+++ src/enamlext/qt/qtable.py
@@ -7,7 +7,6 @@ from enum import Enum, auto
 from typing import Optional, Any, List, Tuple, NamedTuple, Collection, Set, Iterable
 
 # Constants
-from PyQt5.QtWidgets import qApp
 
 from enamlext.qt.table.column import Column, Alignment, AUTO_ALIGN
 from enamlext.qt.table.defs import CellStyle
@@ -15,6 +14,8 @@ from enamlext.qt.table.filtering import TableFilters, Filter
 from qtpy.QtCore import QAbstractTableModel, QModelIndex, Qt, QObject, QPoint, Signal, QItemSelection
 from qtpy.QtGui import QContextMenuEvent, QFont, QColor
 from qtpy.QtWidgets import QApplication, QTableView, QMenu, QAction
+
+from enamlext.qt.table.table_context import TableContext
 
 DEFAULT_ROW_HEIGHT = 23
 DEFAULT_FONT_NAME = "Calibri"
@@ -533,47 +534,6 @@ class QTable(QTableView):
             return set()
 
 
-class TableContext:
-    """
-    Lazy evaluation of properties relative to the data request context.
-    The Qt MVC calls the data() in table model, passing the roles...
-    Depending on the role, we will call callbacks in the column to retrieve things like
-    cell style, font style, tooltip, etc...
-    We pass the object of TableContext to the callback, and everything there is
-    calculated on demand... which hopefully makes it more efficient
-    """
-
-    def __init__(self,
-                 model: QTableModel,
-                 index: QModelIndex,
-                 role: int,
-                 column_index: int,
-                 column: Column,
-                 ):
-        self.__model = model
-        self.index = index
-        self.role = role
-        self.column_index = column_index
-        self.column = column
-
-    @property
-    def row_index(self) -> int:
-        return self.index.row()
-
-    @property
-    def item(self) -> Any:
-        return self.__model.get_item_by_index(self.row_index)
-
-    @property
-    def raw_value(self) -> Any:
-        return self.column.get_value(self.item)
-
-    @property
-    def value(self) -> str:
-        """ Returns the displayed value. """
-        return self.column.get_displayed_value(self.item)
-
-
 class MenuActionContext:
     """
     This class represents what gets passed to the context menu actions (ContextMenuAction)
@@ -722,16 +682,6 @@ class QFilterableHeaderView(QHeaderView):
     def filter_callback(self, column: Column, expression: str):
         # TODO: use column index?
         self.filterChanged.emit(column, expression)
-
-
-# Cell style callbacks
-RED = QColor(Qt.red)
-NEGATIVE_NUMBER_CELL_STYLE = CellStyle(color=RED)
-
-
-def get_cell_style_for_negative_numbers(table_context: TableContext) -> CellStyle:
-    if table_context.raw_value < 0:
-        return NEGATIVE_NUMBER_CELL_STYLE
 
 
 def debug_trace():
