+++ src/enamlext/widgets/dataframe.enaml
@@ -3,8 +3,8 @@ from enamlext.widgets import Table
 import pandas as pd
 from atom.api import Value
 
-# TODO: generate_columns_for_dataframe is in the wrong module - refactor!
-from enamlext.qt.qt_dataframe import generate_columns_for_dataframe, DataFrameProxy
+from enamlext.qt.qt_dataframe import DataFrameProxy
+from enamlext.qt.table.column import generate_columns
 
 
 enamldef DataFrame(Table):
@@ -16,7 +16,7 @@ enamldef DataFrame(Table):
     df ::
         if df is not None:
             self.items = DataFrameProxy(df)
-            self.columns = generate_columns_for_dataframe(df)
+            self.columns = generate_columns(self.items)
         else:
             self.items = []
             self.columns = []
